requirement.txt
flask==2.0.1
pandas==1.3.3
scikit-learn==1.0
nltk==3.6.5
psycopg2-binary==2.9.1
python-dotenv==0.19.1
openpyxl==3.0.9